pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hp-ctl"
version = "0.1.0"
description = "HP control - A modular Python project"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [{name = "Jonas Remmert", email = "j.remmert@mailbox.org"}]
keywords = ["hp", "control"]

dependencies = [
    "pyserial",
    "paho-mqtt>=2.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-pyserial>=0.1.0",
    "types-PyYAML",
]

[project.scripts]
hp-ctl = "hp_ctl.main:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["hp_ctl"]

[tool.setuptools.package-data]
hp_ctl = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "unit: unit tests",
]
log_file = "test_report.log"
log_file_level = "DEBUG"
log_level = "DEBUG"
log_format = "%(asctime)s - %(levelname)s - %(name)s - %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"

[tool.coverage.run]
source = ["src/hp_ctl"]
branch = true
omit = ["*/site-packages/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"


